Flatten TOML config and sync with env vars
